Fix build issue and move storage to class under instance
